Integracao/ContaCapitalIntegracaoLegadoEJB/src/test/java/br/com/sicoob/sisbr/cca/legado/negocio/servicos/interceptors/ContaCapitalIntegracaoLegadoInternacionalizacaoInterceptorTest.java
package br.com.sicoob.sisbr.cca.legado.negocio.servicos.interceptors;

import org.junit.Assert;
import org.junit.BeforeClass;
import org.junit.Test;

import br.com.bancoob.negocio.contexto.InformacoesUsuario;

public class ContaCapitalIntegracaoLegadoInternacionalizacaoInterceptorTest {

	@BeforeClass
	public static void setUpClass() {
		InformacoesUsuario.INSTANCIA.set(new InformacoesUsuario());
	}
	
	@Test
	public void deveInstanciarInterceptor() {
		ContaCapitalIntegracaoLegadoInternacionalizacaoInterceptor interceptor = 
				new ContaCapitalIntegracaoLegadoInternacionalizacaoInterceptor();
		Assert.assertNotNull(interceptor);
		Assert.assertNotNull(interceptor.getResourceBundle());
	}
	
}
